+++ .gitignore
@@ -1,3 +1,2 @@
 redac
-redac-acl
 redac-util
